Polish and Cleanup Add Item Dialog Code
