--- Files UWP/AddItem.xaml
@@ -1,37 +1,0 @@
-﻿<Page
-    x:Class="Files.AddItem"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Files"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{StaticResource ApplicationPageBackgroundThemeBrush}" Loaded="Page_Loaded">
-
-    <Grid>
-        <StackPanel Spacing="25" Orientation="Vertical">
-            <Grid x:Name="SubtitleArea">
-                <TextBlock x:Name="Description" Text="Choose a type below for this new item."/>
-            </Grid>
-            <Grid x:Name="SelectionListContent" HorizontalAlignment="Stretch">
-                <StackPanel>
-                    <ListView x:Name="AddItemsListView" IsItemClickEnabled="True" ItemClick="ListView_ItemClick" ItemsSource="{x:Bind local:AddItem.AddItemsList}">
-                        <ListView.ItemTemplate>
-                            <DataTemplate>
-                                <Grid Height="50">
-                                    <StackPanel Orientation="Horizontal" VerticalAlignment="Center">
-                                        <TextBlock FontFamily="Segoe MDL2 Assets" FontSize="32" Text="{Binding Icon}" Margin="0,0,10,0" VerticalAlignment="Center"/>
-                                        <StackPanel>
-                                            <TextBlock Text="{Binding Header}"/>
-                                            <TextBlock Foreground="Gray" Text="{Binding SubHeader}"/>
-                                        </StackPanel>
-                                    </StackPanel>
-                                </Grid>
-                            </DataTemplate>
-                        </ListView.ItemTemplate>
-                    </ListView>
-                </StackPanel>
-            </Grid>
-        </StackPanel>
-    </Grid>
-</Page>
